automate story discovery
