+++ .storybook/config.js
@@ -5,7 +5,8 @@ import 'babel-polyfill'
 setAddon(JSXAddon)
 
 function loadStories() {
-  require('../stories')
+  const req = require.context('../stories', true, /\.stories\.js$/)
+  req.keys().forEach(filename => req(filename))
 }
 
 configure(loadStories, module)
