--- stories/greyVest/index.js
@@ -1,5 +1,0 @@
-import 'babel-polyfill'
-import { storiesOf } from '@storybook/react'
-
-export default () =>
-  storiesOf('Non Search Components|Grey Vest', module).addWithJSX('Refs', require('./refs').default)
