+++ stories/greyVest/index.stories.js
@@ -1,0 +1,4 @@
+import 'babel-polyfill'
+import { storiesOf } from '@storybook/react'
+
+storiesOf('Non Search Components|Grey Vest', module).addWithJSX('Refs', require('./refs').default)
