+++ stories/queryBuilder/internals/index.stories.js
@@ -26,12 +26,11 @@ let root = {
     testType2: {},
   },
 }
-export default () => {
-  require('./operatorMenu').default(parent, root)
-  require('./operator').default(parent, root, DnDDecorator)
-  require('./addPreview').default(parent, root, DnDDecorator)
-  require('./indentable').default(parent, root, DnDDecorator)
-  require('./filterContents').default(parent, root)
-  require('./rule').default(parent, root, DnDDecorator)
-  require('./group').default(parent, root, DnDDecorator)
-}
+
+require('./operatorMenu').default(parent, root)
+require('./operator').default(parent, root, DnDDecorator)
+require('./addPreview').default(parent, root, DnDDecorator)
+require('./indentable').default(parent, root, DnDDecorator)
+require('./filterContents').default(parent, root)
+require('./rule').default(parent, root, DnDDecorator)
+require('./group').default(parent, root, DnDDecorator)
